requirements.txt
google-cloud-speech
pyaudio
wave
requests
json_tricks